+++ test/end-to-end/just-mr/just-mr-mirrors.test.sh
@@ -179,7 +179,7 @@ if [ ! -s "${CONFIG_CPP}" ]; then
     exit 1
 fi
 
-### Using .just-local specifications
+### Using .jst-local specifications
 
 cat > test-repos.json <<EOF
 { "repositories":
@@ -229,7 +229,7 @@ cat > test-repos.json <<EOF
 }
 EOF
 
-cat > just-local.json <<EOF
+cat > jst-local.json <<EOF
 { "local mirrors":
   { "http://non-existent.example.org/dummy.git":
     [ "http://non-existent.example.org/dummy.git"
@@ -249,11 +249,11 @@ EOF
 
 echo "Test 'local mirrors' fetch"
 rm -rf ${BUILDROOT} ${DISTFILES}/*
-${JUST_MR_CPP} -C test-repos.json --norc --local-build-root ${BUILDROOT} --checkout-locations just-local.json -j 32 fetch --all -o ${DISTFILES}
+${JUST_MR_CPP} -C test-repos.json --norc --local-build-root ${BUILDROOT} --checkout-locations jst-local.json -j 32 fetch --all -o ${DISTFILES}
 
 echo "Test 'local mirrors' setup"
 rm -rf ${BUILDROOT} ${DISTFILES}/*
-CONFIG_CPP=$(${JUST_MR_CPP} -C test-repos.json --norc --local-build-root ${BUILDROOT} --checkout-locations just-local.json -j 32 setup --all)
+CONFIG_CPP=$(${JUST_MR_CPP} -C test-repos.json --norc --local-build-root ${BUILDROOT} --checkout-locations jst-local.json -j 32 setup --all)
 if [ ! -s "${CONFIG_CPP}" ]; then
     exit 1
 fi
@@ -277,7 +277,7 @@ if [ -z "${state_port_num}" ]; then
     exit 1
 fi
 
-### Use .just-local with 'preferred hostnames' favoring host '127.0.0.1' over
+### Use .jst-local with 'preferred hostnames' favoring host '127.0.0.1' over
 ### 'localhost' aka the "state server".
 
 cat > test-repos.json <<EOF
@@ -330,7 +330,7 @@ cat > test-repos.json <<EOF
 }
 EOF
 
-cat > just-local.json <<EOF
+cat > jst-local.json <<EOF
 { "local mirrors":
   {"http://localhost:${state_port_num}/dummy.git": ["${GIT_ROOT}"]}
 , "preferred hostnames": ["127.0.0.1", "localhost"]
@@ -339,7 +339,7 @@ EOF
 
 echo "Test 'preferred hostnames' fetch"
 rm -rf ${BUILDROOT} ${DISTFILES}/*
-${JUST_MR_CPP} -C test-repos.json --norc --local-build-root ${BUILDROOT} --checkout-locations just-local.json -j 32 fetch --all -o ${DISTFILES}
+${JUST_MR_CPP} -C test-repos.json --norc --local-build-root ${BUILDROOT} --checkout-locations jst-local.json -j 32 fetch --all -o ${DISTFILES}
 if [ -f "${STATE_SERVER_ROOT}/access" ]; then
     cat "${STATE_SERVER_ROOT}/access"
     exit 1